requirements.txt
faker==20.1.0
kafka-python==2.0.2
elasticsearch==8.11.0
